# kipartbase/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "Kipartman api specifications"
  version: "1.0.0"
  title: "Kipartman"
  termsOfService: "http://swagger.io/terms/"
  contact:
    name: "Sebastien Besombes"
    url: "https://github.com/turdusmerula/kipartman"
    email: "--"
  license:
    name: "GPL V3"
    url: "https://raw.githubusercontent.com/turdusmerula/kipartman/master/LICENSE"
host: "localhost:8200"
basePath: "/api"
schemes:
- "http"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /version/synchronize:
    post:
      description: "Get synchronization status of a fileset"
      operationId: "synchronize_versioned_files"
      parameters:
      - in: "body"
        name: "files"
        description: "File list to test synchronization"
        required: true
        schema:
          type: "array"
          items:
            $ref: "#/definitions/VersionedFile"
      - name: "root_path"
        in: "query"
        description: "Path from which to synchronize"
        required: false
        type: "string"
      - name: "category"
        in: "query"
        description: "Category of files to see"
        required: false
        type: "string"
      responses:
        200:
          description: "Synchronization state"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/VersionedFile"
        default:
          description: "unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
  /version/commit:
    post:
      description: "Commit a fileset"
      operationId: "commit_versioned_files"
      parameters:
      - in: "body"
        name: "files"
        description: "File list to commit"
        required: true
        schema:
          type: "array"
          items:
            $ref: "#/definitions/VersionedFile"
      - name: "force"
        in: "query"
        description: "Force commit"
        required: false
        type: "boolean"
      responses:
        200:
          description: "Commit result"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/VersionedFile"
        default:
          description: "unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
  /version/update:
    post:
      description: "Update a fileset"
      operationId: "update_versioned_files"
      parameters:
      - in: "body"
        name: "files"
        description: "File list to update"
        required: true
        schema:
          type: "array"
          items:
            $ref: "#/definitions/VersionedFile"
      - name: "force"
        in: "query"
        description: "Force update"
        required: false
        type: "boolean"
      responses:
        200:
          description: "Update result"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/VersionedFile"
        default:
          description: "unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
...
definitions:
...
  Error:
    type: "object"
    required:
    - "code"
    - "message"
    properties:
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"
...
  VersionedFile:
    allOf:
    - $ref: "#/definitions/VersionedFileData"
    - type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        folder:
          type: "boolean"
...
  VersionedFileData:
    type: "object"
    properties:
      source_path:
        type: "string"
      storage_path:
        type: "string"
      md5:
        type: "string"
      version:
        type: "integer"
        format: "int64"
      state:
        type: "string"
      updated:
        type: "string"
        format: "date-time"
      content:
        type: "string"
      category:
        type: "string"
      metadata:
        type: "string"
...